refactor: align ports with other samples, use existing test functions
