+++ policy/policy-01-policy-enforcement/policy-enforcement-consumer/config-us.properties
@@ -1,11 +1,12 @@
-web.http.port=9191
+web.http.port=29191
 web.http.path=/api
-web.http.management.port=9192
+web.http.management.port=29193
 web.http.management.path=/management
-web.http.protocol.port=9292
+web.http.protocol.port=29194
 web.http.protocol.path=/protocol
+
 edc.api.auth.key=password
-edc.dsp.callback.address=http://localhost:9292/protocol
+edc.dsp.callback.address=http://localhost:29194/protocol
 edc.participant.id=consumer
 edc.ids.id=urn:connector:consumer
 edc.mock.region=us
