+++ policy/policy-01-policy-enforcement/policy-enforcement-provider/config.properties
@@ -1,14 +1,10 @@
-web.http.port=8181
+web.http.port=19191
 web.http.path=/api
-web.http.management.port=8182
+web.http.management.port=19193
 web.http.management.path=/management
-web.http.protocol.port=8282
+web.http.protocol.port=19194
 web.http.protocol.path=/protocol
-web.http.control.port=8283
-web.http.control.path=/control
-edc.samples.policy-01.asset.path=/path/to/file
-edc.dsp.callback.address=http://localhost:8282/protocol
+
+edc.dsp.callback.address=http://localhost:19194/protocol
 edc.participant.id=provider
 edc.ids.id=urn:connector:provider
-edc.control.endpoint=http://localhost:8283/control
-edc.samples.policy-01.constraint.location.region=eu
