+++ policy/policy-01-policy-enforcement/resources/catalog-request.json
@@ -2,6 +2,6 @@
   "@context": {
     "@vocab": "https://w3id.org/edc/v0.0.1/ns/"
   },
-  "counterPartyAddress": "http://localhost:8282/protocol",
+  "counterPartyAddress": "http://localhost:19194/protocol",
   "protocol": "dataspace-protocol-http"
 }
