+++ system-tests/src/test/java/org/eclipse/edc/samples/common/NegotiationCommon.java
@@ -84,6 +84,15 @@ public class NegotiationCommon {
                 .pollInterval(POLL_INTERVAL)
                 .until(() -> get(url, CONTRACT_AGREEMENT_ID), Objects::nonNull);
     }
+    
+    public static void checkContractNegotiationState(String contractNegotiationId, String expectedState) {
+        var url = PrerequisitesCommon.CONSUMER_MANAGEMENT_URL + V2_CONTRACT_NEGOTIATIONS_PATH + contractNegotiationId;
+        
+        await()
+                .atMost(TIMEOUT)
+                .pollInterval(POLL_INTERVAL)
+                .until(() -> get(url, "state"), state -> state.equals(expectedState));
+    }
 
     public static String runNegotiation() {
         createAsset();
