+++ system-tests/src/test/java/org/eclipse/edc/samples/policy/Policy01BasicFinalizedTest.java
@@ -22,15 +22,16 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import java.util.Map;
 
-import static org.eclipse.edc.samples.common.PolicyCommon.getFileFromRelativePath;
-import static org.eclipse.edc.samples.common.PolicyCommon.initiateContractNegotiation;
-import static org.eclipse.edc.samples.common.PolicyCommon.lookUpContractAgreementState;
+import static org.eclipse.edc.samples.common.FileTransferCommon.getFileFromRelativePath;
+import static org.eclipse.edc.samples.common.NegotiationCommon.checkContractNegotiationState;
+import static org.eclipse.edc.samples.common.NegotiationCommon.negotiateContract;
 
 @EndToEndTest
 class Policy01BasicFinalizedTest {
     
-    static final String PROVIDER_CONFIG_PROPERTIES_FILE_PATH = "policy/policy-01-policy-enforcement/policy-enforcement-provider/config.properties";
-    static final String CONSUMER_CONFIG_PROPERTIES_FILE_PATH = "policy/policy-01-policy-enforcement/policy-enforcement-consumer/config-eu.properties";
+    private static final String CONTRACT_OFFER_FILE_PATH = "policy/policy-01-policy-enforcement/resources/contractoffer.json";
+    private static final String PROVIDER_CONFIG_PROPERTIES_FILE_PATH = "policy/policy-01-policy-enforcement/policy-enforcement-provider/config.properties";
+    private static final String CONSUMER_CONFIG_PROPERTIES_FILE_PATH = "policy/policy-01-policy-enforcement/policy-enforcement-consumer/config-eu.properties";
     
     @RegisterExtension
     static EdcRuntimeExtension provider = new EdcRuntimeExtension(":policy:policy-01-policy-enforcement:policy-enforcement-provider",
@@ -42,7 +43,7 @@ class Policy01BasicFinalizedTest {
     
     @Test
     void runSampleSteps() {
-        var negotiationId = initiateContractNegotiation();
-        lookUpContractAgreementState(negotiationId, "FINALIZED");
+        var negotiationId = negotiateContract(CONTRACT_OFFER_FILE_PATH, "");
+        checkContractNegotiationState(negotiationId, "FINALIZED");
     }
 }
