+++ system-tests/src/test/java/org/eclipse/edc/samples/policy/Policy01BasicTerminatedTest.java
@@ -21,13 +21,14 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import java.util.Map;
 
-import static org.eclipse.edc.samples.common.PolicyCommon.getFileFromRelativePath;
-import static org.eclipse.edc.samples.common.PolicyCommon.initiateContractNegotiation;
-import static org.eclipse.edc.samples.common.PolicyCommon.lookUpContractAgreementState;
+import static org.eclipse.edc.samples.common.FileTransferCommon.getFileFromRelativePath;
+import static org.eclipse.edc.samples.common.NegotiationCommon.checkContractNegotiationState;
+import static org.eclipse.edc.samples.common.NegotiationCommon.negotiateContract;
 
 @EndToEndTest
 class Policy01BasicTerminatedTest {
     
+    private static final String CONTRACT_OFFER_FILE_PATH = "policy/policy-01-policy-enforcement/resources/contractoffer.json";
     static final String PROVIDER_CONFIG_PROPERTIES_FILE_PATH = "policy/policy-01-policy-enforcement/policy-enforcement-provider/config.properties";
     static final String CONSUMER_CONFIG_PROPERTIES_FILE_PATH = "policy/policy-01-policy-enforcement/policy-enforcement-consumer/config-us.properties";
     
@@ -41,7 +42,7 @@ class Policy01BasicTerminatedTest {
     
     @Test
     void runSampleSteps() {
-        var negotiationId = initiateContractNegotiation();
-        lookUpContractAgreementState(negotiationId, "TERMINATED");
+        var negotiationId = negotiateContract(CONTRACT_OFFER_FILE_PATH, "");
+        checkContractNegotiationState(negotiationId, "TERMINATED");
     }
 }
